+++ src/main/resources/schema.sql
@@ -20,7 +20,7 @@ CREATE TABLE IF NOT EXISTS legal_customer (
 
 CREATE TABLE IF NOT EXISTS deposit (
     id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    amount DOUBLE NOT NULL,
+    amount NUMERIC NOT NULL,
     customer_id BIGINT NOT NULL,
     FOREIGN KEY (customer_id) REFERENCES customer(id) ON DELETE CASCADE
 );
